+++ rnn/island_speciation_strategy.cxx
@@ -108,7 +108,7 @@ IslandSpeciationStrategy::IslandSpeciationStrategy(
     auto make_filled_island = [](int32_t id, RNN_Genome *seed_genome, int32_t size, int32_t nmutations, function<void (RNN_Genome*)> &modify) {
         vector<RNN_Genome*> genomes;
         genomes.reserve(size);
-        for (int i = 0 ; i < size ; i += 1) {
+        for (int32_t i = 0 ; i < size ; i += 1) {
             RNN_Genome *clone = seed_genome->copy();
             modify(clone);
             clone->set_generation_id(0);
@@ -118,7 +118,7 @@ IslandSpeciationStrategy::IslandSpeciationStrategy(
         return new Island(id, genomes);
     };
 
-    for (int i = 0 ; i < number_of_islands; i += 1)
+    for (int32_t i = 0 ; i < number_of_islands; i += 1)
         islands.push_back(make_filled_island(i, _seed_genome, max_island_size, repopulation_mutations, modify));
 
     //set the generation id for the initial minimal genome
@@ -270,8 +270,8 @@ vector<int32_t> IslandSpeciationStrategy::rank_islands() {
         }
     }
 
-    for (int32_t i = 0; i < island_rank.size() - 1; i++)   {
-        for (int32_t j = 0; j < island_rank.size() - i - 1; j++)  {
+    for (int32_t i = 0; i < (int32_t)island_rank.size() - 1; i++)   {
+        for (int32_t j = 0; j < (int32_t)island_rank.size() - i - 1; j++)  {
             fitness_j1 = islands[island_rank[j]]->get_best_fitness();
             fitness_j2 = islands[island_rank[j+1]]->get_best_fitness();
             if (fitness_j1 < fitness_j2) {
@@ -282,7 +282,7 @@ vector<int32_t> IslandSpeciationStrategy::rank_islands() {
         }
     }
     Log::info("island rank: \n");
-    for (int32_t i = 0; i< island_rank.size(); i++){
+    for (int32_t i = 0; i< (int32_t)island_rank.size(); i++){
         Log::info("island: %d fitness %f \n", island_rank[i], islands[island_rank[i]]->get_best_fitness());
     }
     return island_rank;
@@ -451,7 +451,7 @@ RNN_Genome* IslandSpeciationStrategy::generate_genome(uniform_real_distribution<
 
         //set the island for the genome and increment to the next island
         generation_island++;
-        if (generation_island >= (signed) islands.size()) generation_island = 0;
+        if (generation_island >= (int32_t)islands.size()) generation_island = 0;
         islands[generation_island] -> set_latest_generation_id(generated_genomes);
 
     } else {
@@ -625,7 +625,7 @@ RNN_Genome* IslandSpeciationStrategy::parents_repopulation(string method, unifor
 
 void IslandSpeciationStrategy::fill_island(int32_t best_island_id, function<void (int32_t, RNN_Genome*)> &mutate){
     vector<RNN_Genome*>best_island = islands[best_island_id]->get_genomes();
-    for (uint32_t i = 0; i < best_island.size(); i++){
+    for (int32_t i = 0; i < (int32_t)best_island.size(); i++){
         // copy the genome from the best island
         RNN_Genome *copy = best_island[i]->copy();
         generated_genomes++;
@@ -646,7 +646,7 @@ RNN_Genome* IslandSpeciationStrategy::get_global_best_genome(){
 }
 
 void IslandSpeciationStrategy::set_erased_islands_status() {
-    for (int i = 0; i < islands.size(); i++) {
+    for (int32_t i = 0; i < (int32_t)islands.size(); i++) {
         if (islands[i] -> get_erase_again_num() > 0) {
             islands[i] -> set_erase_again_num();
             Log::info("Island %d can be removed in %d rounds.\n", i, islands[i] -> get_erase_again_num());
